tests/wpt/css-tests/css-regions-1_dev/html/regions-keyboard-events-006.htm

<!DOCTYPE html>
<html><head>
        <title>CSS Regions: onkeydown and onkeyup events using Delete key inside region and outside region</title>
        <link href="mailto:dalcala@adobe.com" rel="author" title="David Alcala">
        <link href="mailto:rhauck@adobe.com" rel="reviewer" title="Rebecca Hauck"><!-- 07-19-2013 -->
        <link href="http://www.w3.org/TR/css3-regions/#the-flow-into-property" rel="help">
        <link href="http://www.w3.org/TR/css3-regions/#flow-from" rel="help">
        <link href="http://www.w3.org/TR/css3-regions/#relation-to-document-events" rel="help">
        <meta content="CSS regions module does not alter the normal processing of events in the document
         tree. Therefore, onkeydown and onkeyup events inside and outside a named flow should be raised when the Delete
         key is pressed and released." name="assert">
        <meta content="dom interact" name="flags">
        <link href="support/css/regions-keyboard.css" type="text/css" rel="stylesheet">
        <link href="/resources/testharness.css" type="text/css" rel="stylesheet">
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script src="support/js/regions-keyboard-events-test-helper.js"></script>
    </head>
    <body>
        <ol id="msg">
            <li>Highlight all of the text inside in the text field in the black square</li>
            <li>Press the Delete key to delete the contents of the text field.</li>
            <li>Highlight all of the text inside in the text field in the orange square</li>
            <li>Press the Delete key to delete the contents of the the text field.</li>
        </ol>
        <div id="content">
            <div class="block"></div>
            <div class="block"></div>
            <div class="block" id="target-block">
                <form id="my-form">
                    <input type="text" id="textfield-inside" value="test1X" size="10">
                    <input type="button" id="button-inside" value="Button">
                </form>
            </div>
        </div>
        <div class="region"></div>
        <div class="region"></div>
        <div class="region"></div>
        <br>
        <div id="outside-region">This div is outside the region
            <form id="form-outside">
                <input type="text" id="textfield-outside" value="test2Y" size="10">
                <input type="button" id="button-outside" value="Button">
            </form>
        </div>
        <div id="log"></div>
    
    <script type="text/javascript">

        if(getLeftPosition("target-block") !== 0) {

            var test1 = async_test("onkeydown event fired in textfield inside region", {timeout: testTimeout});
            test1.step(function () {
                var testTarget = document.getElementById("textfield-inside");
                testTarget.onkeydown = test1.step_func(function (evt) {

                    /* Verify the target inside the region has the correct ID */
                    test(function() {
                        assert_equals(evt.target.id, "textfield-inside");
                    }, "Target inside region has the correct ID");

                    test1.done();
                });
            });

            var test2 = async_test("onkeyup event fired in textfield inside region", {timeout: testTimeout});
            test2.step(function () {
                var testTarget = document.getElementById("textfield-inside");
                testTarget.onkeyup = test2.step_func(function (evt) {

                    /* Verify the textfield inside the region has the correct value */
                    test(function() {
                        assert_equals(evt.target.value, "");
                    }, "Textfield inside region has the correct value");

                    test2.done();
                });
                pressDeleteKey("textfield-inside");
            });

            var test3 = async_test("onkeydown event fired in textfield outside region", {timeout: testTimeout});
            test3.step(function () {
                var testTarget = document.getElementById("outside-region");
                testTarget.onkeydown = test3.step_func(function (evt) {

                    /* Verify the target outside the region has the correct ID */
                    test(function () {
                        assert_equals(evt.target.id, "textfield-outside");
                    }, "Target outside region has the correct ID");

                    test3.done();
                });
            });

            var test4 = async_test("onkeyup event fired in textfield outside region", {timeout: testTimeout});
            test4.step( function() {
                var testTarget = document.getElementById("outside-region");
                testTarget.onkeyup = test4.step_func(function (evt) {

                    /* Verify the textfield inside the region has the correct value */
                    test(function () {
                        assert_equals(evt.target.value, "");
                    }, "Textfield outside region has the correct value");

                    test4.done();
                });
                pressDeleteKey("textfield-outside");
            });
            completionCallback();

        } //end of getLeftPosition test
        else {
            test(function () {
                assert_true(false);
            }, "Regions are not enabled");
        }

    </script>

</body></html>
